api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
azure-core==1.32.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.25.1
boto3==1.34.11
botocore==1.34.11
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.2
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
idna==3.6
isodate==0.7.2
jmespath==1.0.1
msal==1.32.0
msal-extensions==1.3.1
pillow==10.2.0
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.10.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.32.3
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.24.0.post1
